+++ Prefabs/droksty.meta
@@ -1,6 +1,7 @@
 fileFormatVersion: 2
-guid: eda15214caff39a45a183f141fc0b87d
-PrefabImporter:
+guid: 1b21d6342e01c654c84b6fa1cf433b87
+folderAsset: yes
+DefaultImporter:
   externalObjects: {}
   userData: 
   assetBundleName:
