+++ Prefabs/droksty/ArtifactPrototype.prefab
