+++ Prefabs/droksty/ArtifactPrototype.prefab.meta
