+++ Prefabs/droksty/CollectiblePrototype.prefab
