+++ Prefabs/droksty/CollectiblePrototype.prefab.meta
