+++ Prefabs/droksty/Level2 Prefabs.meta
@@ -1,5 +1,5 @@
 fileFormatVersion: 2
-guid: a9aa00d3b6db9d847945322b4c0264d3
+guid: 6a662f79dc9d3b84ba71cd7f346d81d8
 folderAsset: yes
 DefaultImporter:
   externalObjects: {}
