+++ Prefabs/droksty/Level2 Prefabs/Exit.prefab.meta
