+++ Prefabs/droksty/Level2 Prefabs/anchor.prefab.meta
