+++ Prefabs/droksty/Player.prefab
@@ -51,7 +51,7 @@ MeshRenderer:
   m_Enabled: 1
   m_CastShadows: 1
   m_ReceiveShadows: 1
-  m_DynamicOccludee: 1
+  m_DynamicOccludee: 0
   m_MotionVectors: 1
   m_LightProbeUsage: 1
   m_ReflectionProbeUsage: 1
@@ -92,7 +92,7 @@ CapsuleCollider:
   m_Material: {fileID: 0}
   m_IsTrigger: 0
   m_Enabled: 1
-  m_Radius: 0.5
+  m_Radius: 0.4
   m_Height: 2
   m_Direction: 1
   m_Center: {x: 0, y: 0, z: 0}
@@ -142,7 +142,7 @@ CharacterController:
   m_Enabled: 1
   serializedVersion: 2
   m_Height: 2
-  m_Radius: 0.5
+  m_Radius: 0.4
   m_SlopeLimit: 40
   m_StepOffset: 0.3
   m_SkinWidth: 0.08
@@ -164,7 +164,9 @@ MonoBehaviour:
   mouseSensitivity: 3
   controller: {fileID: 4792614541385505612}
   walkSpeed: 6
-  runSpeed: 16
+  runSpeed: 15
+  gravityForce: 11
+  jumpForce: 4.5
 --- !u!1 &4792614541413482049
 GameObject:
   m_ObjectHideFlags: 0
