+++ Prefabs/ziggy(lvl2)/cross tombstone.prefab
