+++ Prefabs/ziggy(lvl2)/diamond grave.prefab.meta
