+++ Prefabs/ziggy(lvl2)/simple grave.prefab
