+++ Prefabs/ziggy(lvl2)/wooden cross grave.prefab
