+++ Scripts/droksty/CollectiblePrototypeScript.cs
