+++ Scripts/droksty/CollectiblePrototypeScript.cs.meta
