+++ Scripts/droksty/GameManager.cs
